refactor: move DemoClass to factories
