+++ search/tests/test_course_discovery.py
@@ -3,7 +3,6 @@
 # error, but they do get used when included as part of the override_settings
 """ Tests for search functionalty """
 
-import copy
 import time
 from datetime import datetime
 import ddt
@@ -16,63 +15,10 @@ from meilisearch.errors import MeilisearchApiError
 
 from search.api import course_discovery_search, NoSearchEngineError
 from search.elastic import ElasticSearchEngine
+from search.tests.factories import DemoCourse
 from search.tests.utils import SearcherMixin, TEST_INDEX_NAME
 from search.meilisearch import get_meilisearch_client, create_indexes
 from .mock_search_engine import MockSearchEngine
-
-
-class DemoCourse:
-    """ Class for dispensing demo courses """
-    DEMO_COURSE_ID = "edX/DemoX/Demo_Course"
-    DEMO_COURSE = {
-        "start": datetime(2014, 2, 1),
-        "number": "DemoX",
-        "content": {
-            "short_description": "Short description",
-            "overview": "Long overview page",
-            "display_name": "edX Demonstration Course",
-            "number": "DemoX"
-        },
-        "course": "edX/DemoX/Demo_Course",
-        "image_url": "/c4x/edX/DemoX/asset/images_course_image.jpg",
-        "effort": "5:30",
-        "id": DEMO_COURSE_ID,
-        "enrollment_start": datetime(2014, 1, 1),
-    }
-
-    demo_course_count = 0
-
-    @classmethod
-    def get(cls, update_dict=None, remove_fields=None):
-        """ get a new demo course """
-        cls.demo_course_count += 1
-        course_copy = copy.deepcopy(cls.DEMO_COURSE)
-        if update_dict:
-            if "content" in update_dict:
-                course_copy["content"].update(update_dict["content"])
-                del update_dict["content"]
-            course_copy.update(update_dict)
-        course_copy.update({"id": "{}_{}".format(course_copy["id"], cls.demo_course_count)})
-        if remove_fields:
-            for remove_field in remove_fields:
-                if remove_field in course_copy:
-                    del course_copy[remove_field]
-        return course_copy
-
-    @classmethod
-    def reset_count(cls):
-        """ go back to zero """
-        cls.demo_course_count = 0
-
-    @staticmethod
-    def index(searcher, course_info):
-        """ Adds course info dictionary to the index """
-        searcher.index(sources=course_info)
-
-    @classmethod
-    def get_and_index(cls, searcher, update_dict=None, remove_fields=None):
-        """ Adds course info dictionary to the index """
-        cls.index(searcher, [cls.get(update_dict, remove_fields)])
 
 
 @override_settings(SEARCH_ENGINE="search.tests.mock_search_engine.MockSearchEngine")
